moves selectors to selectors Folder
