+++ app/modules/auth/reducer.ts
@@ -29,10 +29,3 @@ export const authReducer: AppReducer<IAuthState> = (
 
   return state;
 };
-
-export const selectIsAuthorized = (state: IAuthState): boolean => !!(state.jwt && state.user);
-export const selectUserEmail = (state: IAuthState): string | undefined =>
-  state.user && (state.user.unverifiedEmail || state.user.verifiedEmail);
-export const selectUser = (state: IAuthState): IUser | undefined => state.user;
-export const selectBackupCodesVerified = (state: IAuthState): boolean | undefined =>
-  state.user && state.user.backupCodesVerified;
